Progress bar fixes
* Instead of converting the time variables to int, just print them as
doubles without any decimals.
* Doing this revealed that occasionally the 'secs' value would
overflow as an int, and become NaN as double. So add a check for this
before printing, as well as make sure secs is larger than 0 just in
case.
* Erase the previously written line before we start printing again on
the rolled back line. This fixes the issue where if the output shrank
in size it would still keep some lingering characters to the right.