Add ability to compare file modification date
Rather than depending solely on the filesize (which could potentially
be the same but with the file still changed), we now attempt to compare
the file modification date first if available. If not we fall back on
comparing file sizes.

Also fix a bug with removing php/script arguments from urls when
determining file names, where it would remove one letter too much.