Fix issue with downloads always reporting failures even on success
I accidentally removed the part where we save the result of the
download into "res", which we use to check if the download was
successful or not. So it was always false.

Also replaced "curler" with "argv[0]" in the help message, in case
someone would compile the application with another name.