Fix issue with setting filetime, and other small tweaks
Not sure how I missed this, but when doing the big refactor the other
day I forgot to remove the part in the download function where we used
to check the filetime. We now check the filetime in the get_headers()
function, so the one in the download function overwrote the filetime
which lead to it not being able to set it.