Attempt to more accurately determine generic filename
Before we just checked if the filename existed before determining if
the generic filename was taken or not. This didn't work since we then
append the filetype to the full path, and also wouldn't work unless we
downloaded the file to the current working dir. This attempts to fix
that.