Add ability to get filename from location header
Extract header location in header_callback(), and use it to get a
hopefully more correct filename in find_filename().

Also add a special case for when the location header was found AND the
filetype is ".html", since the location header is always (?) found on
an html page. So setting the filetype to html would be incorrect. From
my testing the correct filetype is present in the location url so we
extract it that way.

There's now also a check for any '&' characters in find_filename() if
we don't find the content-disposition, since similarly to '?' those
are usually used for url arguments. Hopefully that won't bite us.