Fix memory leak with curl_easy_unescape()
I didn't realize that curl_easy_unescape() returns a char*, and that
by assigning it to a std::string it will never be deallocated. Solved
this by assigning it to a temporary char* which then is copied to the
std::string, and the char* is then deallocated using curl_free().