Fix for another time display issue
If there was >1 hour left, and mins was 0, it wouldn't display the
mins part. I don't really like this solution - most likely the same
issue would affect the days - but it will do for now since most
downloads won't take days. But I may have to rethink how we determine
what parts are printed or not.